Fix some exports
