+++ src/components/index.ts
@@ -332,3 +332,6 @@ export { default as WorkPoolStatusSelect } from './WorkPoolStatusSelect.vue'
 export { default as WorkQueueStatusSelect } from './WorkQueueStatusSelect.vue'
 export { default as EventResourceCombobox } from './EventResourceCombobox.vue'
 export { default as EventsCombobox } from './EventsCombobox.vue'
+export { default as WorkspaceEventsTimeline } from './WorkspaceEventsTimeline.vue'
+export { default as EventResourceIconText } from './EventResourceIconText.vue'
+export { default as ResourceLink } from './ResourceLink.vue'
